refactor: move logic to controller file
